+++ src/lib.rs
@@ -46,6 +46,7 @@ pub mod install_recipes;
 pub mod install_selinux_mail;
 pub mod install_server;
 pub mod install_snappymail_prefs;
+pub mod install_snappymail_repo;
 pub mod install_webmail;
 pub mod install_webmail_health;
 pub mod install_webmail_proxy;